+++ PhysPix Sandbox/src/game/cell/Oil.cpp
@@ -8,7 +8,7 @@
 
 namespace cell
 {
-	Oil::Oil(CellGrid* cellGrid, int32_t posX, int32_t posY)
+	Oil::Oil(CellGrid* cellGrid, uint16_t posX, uint16_t posY)
 	{
 		m_CellGrid = cellGrid;
