+++ PhysPix Sandbox/src/game/cell/Smoke.cpp
@@ -2,7 +2,7 @@
 
 namespace cell
 {
-	Smoke::Smoke(CellGrid* cellGrid, int32_t posX, int32_t posY)
+	Smoke::Smoke(CellGrid* cellGrid, uint16_t posX, uint16_t posY)
 	{
 		m_CellGrid = cellGrid;
