+++ PhysPix Sandbox/src/game/cell/Steam.cpp
@@ -2,7 +2,7 @@
 
 namespace cell
 {
-	Steam::Steam(CellGrid* cellGrid, int32_t posX, int32_t posY)
+	Steam::Steam(CellGrid* cellGrid, uint16_t posX, uint16_t posY)
 	{
 		m_CellGrid = cellGrid;
