+++ PhysPix Sandbox/src/game/cell/Steam.h
@@ -8,7 +8,7 @@ namespace cell
 	class Steam : public Cell, public Fluid
 	{
 	public:
-		Steam(CellGrid* cellGrid, int32_t posX, int32_t posY);
+		Steam(CellGrid* cellGrid, uint16_t posX, uint16_t posY);
 		~Steam();
 
 		void OnTick() override;
